+++ superset-frontend/src/components/Skeleton/Skeleton.test.tsx
@@ -18,7 +18,7 @@
  */
 import { render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { Skeleton } from 'antd';
+import { Skeleton } from '.';
 
 describe('Skeleton Component', () => {
   test('renders skeleton', () => {
