+++ superset-frontend/src/components/Space/index.tsx
@@ -16,12 +16,4 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
-/*
- * Re-exporting of components in src/components to facilitate
- * their imports by other components.
- * E.g. import { Select } from 'src/components'
- */
-
-export { Space } from 'antd-v5';
-export type { SpaceProps } from 'antd-v5/es/space';
+export { Space, type SpaceProps } from 'antd-v5';
