+++ superset-frontend/src/components/Steps/Steps.stories.tsx
@@ -17,7 +17,7 @@
  * under the License.
  */
 import { Steps as AntdSteps } from 'antd-v5';
-import { Steps, StepsProps } from '.';
+import { Steps, type StepsProps } from '.';
 
 export default {
   title: 'Components/Steps',
