+++ superset-frontend/src/components/Switch/Switch.stories.tsx
@@ -17,7 +17,7 @@
  * under the License.
  */
 import { useArgs } from '@storybook/preview-api';
-import { Switch, SwitchProps } from '.';
+import { Switch, type SwitchProps } from '.';
 
 export default {
   title: 'Components/Switch',
