+++ superset-frontend/src/components/TreeSelect/TreeSelect.test.tsx
@@ -18,7 +18,7 @@
  */
 import { render, fireEvent, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import TreeSelect from '.';
+import { TreeSelect } from '.';
 
 const treeData = [
   {
