+++ superset-frontend/src/explore/components/EmbedCodeContent.jsx
@@ -18,12 +18,10 @@
  */
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { css, t } from '@superset-ui/core';
-import { Input, CopyToClipboard } from 'src/components';
+import { Input, CopyToClipboard, Space, Typography } from 'src/components';
 import { URL_PARAMS } from 'src/constants';
 import { getChartPermalink } from 'src/utils/urlUtils';
-import { Typography } from 'src/components/Typography';
 import { Icons } from 'src/components/Icons';
-import { Space } from 'src/components/Space';
 
 const EmbedCodeContent = ({ formData, addDangerToast }) => {
   const [height, setHeight] = useState('400');
